give GetChangedGradients its own file
